requirements.txt
Click==7.0
dominate==2.3.5
Flask==1.0.2
Flask-Bootstrap==3.3.7.1
Flask-WTF==0.14.2
itsdangerous==1.1.0
Jinja2==2.10
MarkupSafe==1.1.0
numpy==1.16.0
opencv-python==4.0.0.21
Pillow==5.4.1
pytesseract==0.2.6
python-dotenv==0.10.1
visitor==0.1.3
Werkzeug==0.14.1
WTForms==2.2.1
